Treat only FORCE prerequisite specially
Other phony prerequisites of `cmd`-based rules are now still treated
normally (i.e., always out-of-date) rather than being filtered out as
well.

It appears that KBuild treats all phony prerequisites that way because
it came to depend on a bug in old versions of Make, and didn't have
order-only dependencies as a workaround once the bug was fixed. We can
use order-only dependencies when we want them, so try to keep the magic
to a minimum.